🚀 Create separate App.windows for windows.
It does not matter if native modules (that are not supported on windows yet, i. e. they do not contain folder named windows) are installed (in package.json) as long as they are not used in called code.
By naming file SomeComponent.windows.tsx, this file is used preferably for Windows.
Also as a matter of this change, I was able to uncomment import of react-native-gesture-handler that is advised (but not necessary) for react-navigation.