🚀 Add FilePicker.cpp that provides functionality to select an image on Windows. Use MyImage interface over Image.
FilePicker also copies the file to a temporary directory so that it can be later accessed (https://github.com/microsoft/react-native-windows/issues/4255). It is not integrated yet to the JS app, because there were issues with posting the base64 string data to the Google Storage. Google Storage relies on Blob implementation and React Native has none. I tried fixing it with Blob polyfill from react-native-blob-util,, which made the error disappear, but 'StorageException: An unknown error occurred' appeared instead.